+++ python/dalex/dalex/instance_level/_break_down/object.py
@@ -90,7 +90,7 @@ class BreakDown:
         title : str, optional
             Title of the plot (default is "Break Down").
         vertical_spacing : float <0, 1>, optional
-            Ratio of vertical space between the plots (default is 0.2/number of subplots).
+            Ratio of vertical space between the plots (default is 0.2/number of rows).
         show : bool, optional
             True shows the plot; False returns the plotly Figure object that can be
             edited or saved using the `write_image()` method (default is True).
@@ -217,8 +217,8 @@ class BreakDown:
 
         if show:
             fig.show(config={
-                'displaylogo': False,
-                'staticPlot': False,
+                'displaylogo': False, 'staticPlot': False,
+                'toImageButtonOptions': {'height': None, 'width': None, },
                 'modeBarButtonsToRemove': ['sendDataToCloud', 'lasso2d', 'autoScale2d', 'select2d', 'zoom2d', 'pan2d',
                                            'zoomIn2d', 'zoomOut2d', 'resetScale2d', 'toggleSpikelines',
                                            'hoverCompareCartesian', 'hoverClosestCartesian']